requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
mysqlclient==2.1.0
pyserial==3.5
pytz==2022.1
sqlparse==0.4.2
